backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
groq==0.13.0
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
python-multipart==0.0.18
aiofiles==24.1.0
httpx==0.28.1
python-jose==3.3.0
sqlalchemy==2.0.36
duckdb==1.1.3
pyarrow>=14.0.0
psycopg2-binary==2.9.10
pymysql==1.1.1
pyodbc==5.2.0
scipy==1.13.1
scikit-learn==1.5.2
tabulate==0.9.0
